// frontend/cloudfarm_farm/src/app/models/User.ts
export interface User {
    user_id: number;
    first_name: string;
    last_name: string;
    birth_date: Date;
    phone_number: string;
    email_address: string;
    password: string;
    physical_address: string;
    city: string;
    state: string;
    country: string;
    user_status: string;
}